Users Moderation done

+++ apps/backend/src/Gump.Data/Repositories/RecipeRepository.cs
@@ -267,8 +267,11 @@ public partial class RecipeRepository : RepositoryBase<RecipeModel>
 			throw new RestrictedException("Recipe can only be archived, because it has saves");
 		}
 
-		var image = ImageRepository.GetById(recipe.ImageId);
-		ImageRepository.Delete(image.Id);
+		if (recipe.ImageId != 1)
+		{
+			var image = ImageRepository.GetById(recipe.ImageId);
+			ImageRepository.Delete(image.Id);
+		}
 
 		try
 		{
